requirements.txt
torch
pandas
numpy
yfinance
scikit-learn
fastapi
uvicorn
vercel-asgi